[REFACTORING] Move PullRequest to DarcLib.Models
