+++ src/Microsoft.DotNet.Darc/DarcLib/GitHubClient.cs
@@ -300,7 +300,7 @@ public class GitHubClient : RemoteRepoBase, IRemoteGitRepo
     /// </summary>
     /// <param name="pullRequestUrl">Uri of the pull request</param>
     /// <returns>Information on the pull request.</returns>
-    public async Task<PullRequest> GetPullRequestAsync(string pullRequestUrl)
+    public async Task<Models.PullRequest> GetPullRequestAsync(string pullRequestUrl)
     {
         (string owner, string repo, int id) = ParsePullRequestUri(pullRequestUrl);
         Octokit.PullRequest pr = await GetClient(owner, repo).PullRequest.Get(owner, repo, id);
@@ -315,7 +315,7 @@ public class GitHubClient : RemoteRepoBase, IRemoteGitRepo
             status = PrStatus.Open;
         }
 
-        return new PullRequest
+        return new Models.PullRequest
         {
             Title = pr.Title,
             Description = pr.Body,
@@ -332,7 +332,7 @@ public class GitHubClient : RemoteRepoBase, IRemoteGitRepo
     /// </summary>
     /// <param name="repoUri">Repo to create the pull request for.</param>
     /// <param name="pullRequest">Pull request data</param>
-    public async Task<string> CreatePullRequestAsync(string repoUri, PullRequest pullRequest)
+    public async Task<string> CreatePullRequestAsync(string repoUri, Models.PullRequest pullRequest)
     {
         (string owner, string repo) = ParseRepoUri(repoUri);
 
@@ -359,7 +359,7 @@ public class GitHubClient : RemoteRepoBase, IRemoteGitRepo
     /// </summary>
     /// <param name="pullRequestUri">Uri of pull request to update</param>
     /// <param name="pullRequest">Pull request info to update</param>
-    public async Task UpdatePullRequestAsync(string pullRequestUri, PullRequest pullRequest)
+    public async Task UpdatePullRequestAsync(string pullRequestUri, Models.PullRequest pullRequest)
     {
         (string owner, string repo, int id) = ParsePullRequestUri(pullRequestUri);
 
@@ -1232,7 +1232,7 @@ public class GitHubClient : RemoteRepoBase, IRemoteGitRepo
     /// <returns>Async task</returns>
     public async Task DeletePullRequestBranchAsync(string pullRequestUri)
     {
-        PullRequest pr = await GetPullRequestAsync(pullRequestUri);
+        Models.PullRequest pr = await GetPullRequestAsync(pullRequestUri);
         (string owner, string repo, int id) prInfo = ParsePullRequestUri(pullRequestUri);
         await DeleteBranchAsync(prInfo.owner, prInfo.repo, pr.HeadBranch);
     }
